requirements.txt
SQLAlchemy
python-dotenv
psycopg2
grpcio
protobuf
grpcio-tools
fastapi
uvicorn
pydantic 